requirements.txt
altgraph==0.17.2
grpcio==1.45.0
grpcio-tools==1.45.0
protobuf==3.20.1
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyYAML==6.0
six==1.16.0
